VM001 - Refactor logging & debugging into 'utils/'
