+++ backend/src/utils/debug.ts
@@ -1,0 +1,3 @@
+import debugConstructor from "debug";
+
+export const debug = debugConstructor("APP");
